requirements.txt
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
docutils==0.21.2
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
keyring==25.4.1
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.5.0
nh3==0.2.18
packaging==24.1
parse==1.20.2
pkginfo==1.10.0
pluggy==1.5.0
pycparser==2.22
Pygments==2.18.0
pytest==8.3.3
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
requests-wsgi-adapter==0.4.1
rfc3986==2.0.0
rich==13.9.2
SecretStorage==3.3.3
setuptools==75.1.0
tqdm==4.66.5
twine==5.1.1
urllib3==2.2.3
WebOb==1.8.8
wheel==0.44.0
zipp==3.20.2
